# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MapVizit API - Map & Digital Business Card Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [sudo, admin, premium, viewer]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user role or subscription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
...
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get app settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppSettings"
    patch:
      operationId: updateSettings
      tags: [settings]
      summary: Update app settings (sudo only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppSettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppSettings"
...
  /settings/import:
    post:
      operationId: importData
      tags: [settings]
      summary: Import DB backup (sudo only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BackupData"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
components:
  schemas:
...
    User:
      type: object
      required: [id, username, name, role, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        phone:
          type: ["string", "null"]
        telegramId:
          type: ["string", "null"]
        telegramUsername:
          type: ["string", "null"]
        role:
          type: string
          enum: [sudo, admin, premium, viewer]
        isPremium:
          type: boolean
        premiumExpiresAt:
          type: ["string", "null"]
        assignedPointId:
          type: ["integer", "null"]
        createdAt:
          type: string
    UserUpdate:
      type: object
      properties:
        role:
          type: string
          enum: [sudo, admin, premium, viewer]
        isPremium:
          type: boolean
        premiumExpiresAt:
          type: ["string", "null"]
        assignedPointId:
          type: ["integer", "null"]
...
    AppSettings:
      type: object
      required: [premiumEnabled, botToken, backupChatId]
      properties:
        premiumEnabled:
          type: boolean
        botToken:
          type: ["string", "null"]
        backupChatId:
          type: ["string", "null"]
        appName:
          type: ["string", "null"]
    AppSettingsUpdate:
      type: object
      properties:
        premiumEnabled:
          type: boolean
        botToken:
          type: ["string", "null"]
        backupChatId:
          type: ["string", "null"]
        appName:
          type: ["string", "null"]
    BackupData:
      type: object
      required: [version, exportedAt]
      properties:
        version:
          type: string
        exportedAt:
          type: string
        users:
          type: array
          items:
            type: object
        categories:
          type: array
          items:
            type: object
        points:
          type: array
          items:
            type: object
        contacts:
          type: array
          items:
            type: object
        images:
          type: array
          items:
            type: object
        settings:
          type: array
          items:
            type: object
    ImportResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
        imported:
          type: object
          properties:
            users:
              type: integer
            categories:
              type: integer
            points:
              type: integer