# lib/api-zod/src/generated/types/appSettingsUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MapVizit API - Map & Digital Business Card Platform
 * OpenAPI spec version: 0.1.0
 */

export interface AppSettingsUpdate {
  premiumEnabled?: boolean;
  /** @nullable */
  botToken?: string | null;
  /** @nullable */
  backupChatId?: string | null;
  /** @nullable */
  appName?: string | null;
}